move inquiries from constants to prompt
